release 4.0.24

+++ include/Math/ShapeGeneration/Ellipse.h
@@ -1,4 +1,4 @@
-// IDDN FR.001.250001.004.S.X.2019.000.00000
+// IDDN.FR.001.250001.005.S.P.2019.000.00000
 // ULIS is subject to copyright laws and is the legal and intellectual property of Praxinos,Inc
 /*
 *   ULIS
@@ -6,7 +6,6 @@
 * @file         Ellipse.h
 * @author       Thomas Schmitt
 * @brief        This file provides the Ellipse points generation methods
-* @copyright    Copyright 2018-2021 Praxinos, Inc. All Rights Reserved.
 * @license      Please refer to LICENSE.md
 */
 #pragma once
@@ -49,7 +48,7 @@ static inline void GenerateEllipsePoints(
         x++;
     }
 
-    octantStartingIndex = ioEllipsePoints.Size();
+    octantStartingIndex = static_cast< int >( ioEllipsePoints.Size() );
 
     //Octant 2 ------
     x = iA;
@@ -68,7 +67,7 @@ static inline void GenerateEllipsePoints(
         y++;
     }
 
-    octantStartingIndex = ioEllipsePoints.Size();
+    octantStartingIndex = static_cast< int >( ioEllipsePoints.Size() );
 
     //Octant 3 ------
     x = iA;
@@ -88,7 +87,7 @@ static inline void GenerateEllipsePoints(
         y++;
     }
 
-    octantStartingIndex = ioEllipsePoints.Size();
+    octantStartingIndex = static_cast< int >( ioEllipsePoints.Size() );
    
     //Octant 4 ------
     x = 0;
@@ -108,7 +107,7 @@ static inline void GenerateEllipsePoints(
         x++;
     }
 
-    octantStartingIndex = ioEllipsePoints.Size();
+    octantStartingIndex = static_cast< int >( ioEllipsePoints.Size() );
 
     //Octant 5 ------
     x = 0;
@@ -128,7 +127,7 @@ static inline void GenerateEllipsePoints(
         x++;
     }
 
-    octantStartingIndex = ioEllipsePoints.Size();
+    octantStartingIndex = static_cast< int >( ioEllipsePoints.Size() );
 
     //Octant 6 ------
     x = iA;
@@ -148,7 +147,7 @@ static inline void GenerateEllipsePoints(
         y++;
     }
 
-    octantStartingIndex = ioEllipsePoints.Size();
+    octantStartingIndex = static_cast< int >( ioEllipsePoints.Size() );
 
     //Octant 7 ------
     x = iA;
@@ -168,7 +167,7 @@ static inline void GenerateEllipsePoints(
         y++;
     }
 
-    octantStartingIndex = ioEllipsePoints.Size();
+    octantStartingIndex = static_cast< int >( ioEllipsePoints.Size() );
 
     //Octant 8 ------
     x = 0;
